Don't destroy things when removing ZigBee nodes from the network
This makes it easier to remove/rejoin zigbee nodes without losing
all magic and other setup related to a thing.

Comes with the cost of having to manually remove a thing when
removing a thing from the ZigBee network for good. That should be
a bearable cost though, as removing a Thing will still remove the
associated node in one go.